Se mueve navigator de Auth desde App.tsx a index.tsx
